+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Added an optional API key request parameter to skip PoW in faucet (#839).
 - Block producer now emits network note information (#833).
 - Introduced Network Transaction Builder (#840).
+- Added way of executing and proving network transactions (#841).
 
 ## v0.8.2 (2025-05-04)